requirements.txt
matplotlib
nordpool
numpy
pandas
python-dotenv
requests
scikit_learn
scipy
statsmodels
xgboost
yfinance
